Fix "get_working_cache" to work even when objects do not exist
FSSpec provides an `exists(<rpath>)` method that checks whether or not
the file designated by `<rpath>` currently exists.

I'm using this method to integrate PelicanFS with my snakemake Pelican
storage plugin, but without this change, the `exists()` method returns
a fatal error whenever a cache returns a 404.

This is problematic because the method should work to check whether
an object exists, not whether the cache has it.