Forward redirected stdin data to subprocess in Jupyter
When stdin is redirected (e.g., StringIO in Jupyter), read all available data from it and write it to the PTY master_fd. This allows password input to work when stdin is programmatically redirected.

The data is sent during each iteration of the select loop until consumed, then cleared to avoid repeated sends.